finished? refactor
